Fix allow overriding of rollup and rollup config.
