+++ defs.bzl
@@ -24,7 +24,7 @@ load("//internal/node:node.bzl",
 load("//internal/node:node_repositories.bzl",_node_repositories = "node_repositories")
 load("//internal/jasmine_node_test:jasmine_node_test.bzl", _jasmine_node_test = "jasmine_node_test")
 load("//internal/npm_install:npm_install.bzl", _npm_install = "npm_install", _yarn_install = "yarn_install")
-load("//internal/rollup:rollup_bundle.bzl", _rollup_bundle = "rollup_bundle")
+load("//internal/rollup:rollup_bundle.bzl", _rollup_bundle = "rollup_bundle_macro")
 load("//internal/npm_package:npm_package.bzl", _npm_package = "npm_package")
 
 check_bazel_version = _check_bazel_version
