+++ internal/e2e/rollup/bar.js
@@ -1,1 +1,1 @@
-export { name } from './blob.json'
+export const name = 'Alice';
