--- internal/e2e/rollup/blob.json
@@ -1,4 +1,0 @@
-{
-    "name": "Alice",
-    "age": "seven"
-}
