+++ internal/e2e/rollup/golden/bundle-min-debug_golden.js_
