+++ internal/e2e/rollup/golden/bundle-min_golden.js_
