+++ internal/e2e/rollup/golden_override/bundle-min-debug_golden.js_
@@ -1,0 +1,21 @@
+/**
+ * @license A dummy license banner that goes at the top of the file.
+ * This is version v1.2.3
+ */
+var bundle_override = function(exports, some_global_var) {
+    "use strict";
+    console.log("Hello, Alice in Wonderland");
+    var A = function() {
+        function A() {}
+        A.prototype.a = function() {
+            return document.a;
+        };
+        return A;
+    }();
+    function inline_me() {
+        return "abc";
+    }
+    console.error(new A().a(), inline_me(), some_global_var.thing);
+    exports.A = A;
+    return exports;
+}({}, runtime_name_of_global_var);
