+++ internal/e2e/rollup/rollup.config.js
@@ -1,0 +1,29 @@
+// Rollup configuration
+// GENERATED BY Bazel
+
+const rollup = require('rollup');
+const nodeResolve = require('rollup-plugin-node-resolve');
+const sourcemaps = require('rollup-plugin-sourcemaps');
+const {resolveBazel, notResolved, banner, onwarn, moduleDirectory} =
+    require('TMPL_bazel_rollup_conf_path');
+
+module.exports = {
+  resolveBazel,
+  banner,
+  onwarn,
+  output: {
+    format: 'TMPL_output_format',
+    name: 'TMPL_global_name',
+  },
+  plugins: [
+    {resolveId: resolveBazel},
+    nodeResolve({
+      preferBuiltins: false,
+      jsnext: true,
+      module: true,
+      customResolveOptions: {moduleDirectory}
+    }),
+    {resolveId: notResolved},
+    sourcemaps(),
+  ]
+}
